src/requirements.txt
streamlit
python-dotenv
bcrypt
google-generativeai
